refactor(CodeQuality): 高复杂度代码重构与EF迁移目录统一 consolidate-code-quality
## 代码质量改进

### Phase 1: 高复杂度方法重构
- PatientImportExecutor: 消除反射模式，创建ImportResult记录类型
- MedicalCaseRepository.UpdateAsync: 提取4个辅助方法处理处方状态
- ExcelHelper: SetCellValue使用switch表达式，ConvertValue提取5个辅助方法
- MedicalCaseCommandService.SaveAsync: 提取10+辅助方法简化保存逻辑

### Phase 2: EF迁移目录整合
- 统一迁移目录：从Data/Migrations/移动5个迁移到Migrations/
- 更新命名空间：LYBT.Infrastructure.Data.Migrations → LYBT.Infrastructure.Migrations
- 删除冗余目录：Data/Migrations/

### 额外修复
- 添加缺失的DecocteMethodToVisibilityConverter转换器

### 验证结果
- 编译：0错误
- 测试：全部通过

Related: openspec consolidate-code-quality

+++ src/Server/Modules/LYBT.Module.Formula/Services/FormulaService.cs
@@ -128,7 +128,7 @@ namespace LYBT.Module.Formulas.Services
                 {
                     HerbId = h.HerbId,
                     HerbName = h.HerbName,
-                    Dosage = (int)h.Dosage, // decimal → int
+                    Dosage = h.Dosage, // decimal → int
                     Unit = h.Unit,
                     ProcessingMethod = h.ProcessingMethod ?? h.Preparation, // 优先使用ProcessingMethod，回退到Preparation
                     Usage = h.Usage,
@@ -173,7 +173,7 @@ namespace LYBT.Module.Formulas.Services
                     {
                         HerbId = h.HerbId,
                         HerbName = h.HerbName,
-                        Dosage = (int)h.Dosage, // decimal → int
+                        Dosage = h.Dosage, // decimal → int
                         Unit = h.Unit,
                         ProcessingMethod = h.ProcessingMethod ?? h.Preparation, // 优先使用ProcessingMethod
                         Usage = h.Usage,
@@ -365,7 +365,7 @@ namespace LYBT.Module.Formulas.Services
                             HerbName = herbDto.HerbName,
                             OriginalHerbName = herbDto.HerbName, // 保存原始名称
                             IsValidated = matchedHerb != null, // 成功匹配则标记为已验证
-                            Dosage = (int)herbDto.Dosage, // DTO是decimal，实体是int
+                            Dosage = herbDto.Dosage, // DTO是decimal，实体是int
                             Unit = herbDto.Unit ?? string.Empty,
                             Usage = herbDto.Usage,
                             ProcessingMethod = herbDto.Preparation // DTO的Preparation映射到ProcessingMethod
